requirements.txt
asgiref==3.5.2
Django==3.2.13
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-jsonfield==1.4.1
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-tenants==3.4.2
gunicorn==20.1.0
jsonfield==3.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
swapper==1.3.0
